remove executable and split up a bit
