+++ Cargo.toml
@@ -15,7 +15,5 @@ categories = ["command-line-utilities", "cryptography"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-clap = { version = "4.5.4", features = ["derive"] }
-once_cell = "1.19.0"
 rand = "0.8.5"
 thiserror = "1.0.61"
