--- src/main.rs
@@ -1,13 +1,0 @@
-use std::process::exit;
-
-use pants_gen::cli::CliArgs;
-
-fn main() {
-    let password = CliArgs::run();
-    if let Some(p) = password {
-        println!("{}", p);
-    } else {
-        println!("Contraints couldn't be met, try again after adjusting the spec");
-        exit(1)
-    }
-}
